+++ tests/test_formatter/test_reformat_markdown_safe.py
@@ -1,13 +1,13 @@
 from tglogger import formatter
 
 
-def test_reformat_markdown_safe_underscore():
+def test__reformat_markdown_safe_underscore():
     example = "foo_bar_baz"
-    reformatted = formatter.reformat_markdown_safe(example)
-    assert reformatted == "foo\_bar\_baz"
+    reformatted = formatter._reformat_markdown_safe(example)
+    assert reformatted == "foo\_bar\_baz"  # noqa
 
 
-def test_reformat_markdown_safe_asterisk():
+def test__reformat_markdown_safe_asterisk():
     example = "foo*bar*baz"
-    reformatted = formatter.reformat_markdown_safe(example)
-    assert reformatted == "foo\*bar\*baz"
+    reformatted = formatter._reformat_markdown_safe(example)
+    assert reformatted == "foo\*bar\*baz"  # noqa