+++ tglogger/__init__.py
@@ -1,2 +1,7 @@
-__version__ = "0.1.1a2"
+"""
+`tglogger` is a third-party module containing logging utilities for logging on Telegram.
+It contains logging handlers and formatters and utilities to process logs.
+"""
+
+__version__ = "0.1.1a3"
 __author__ = "Eray Erdin"