+++ tglogger/formatter.py
@@ -1,20 +1,28 @@
-# todo - 2 class doc
+"""
+This module contains a formatter for logging with Telegram.
+"""
 
 import logging
 import os
 from textwrap import dedent
 
 
-def reformat_markdown_safe(text: str) -> str:
-    # todo - 2 function doc
+def _reformat_markdown_safe(text: str) -> str:
+    """
+    Sanitizes the provided text of special markdown characters. `_` and `*` characters
+    are sanitized.
+    """
 
-    text = text.replace("_", "\_")
-    text = text.replace("*", "\*")
+    text = text.replace("_", "\_")  # noqa
+    text = text.replace("*", "\*")  # noqa
     return text
 
 
 class TelegramFormatter(logging.Formatter):
-    # todo 3 - class doc
+    """
+    A `logging.Formatter` implementation for sending log records with Telegram.
+    """
+
     _TEMPLATE = dedent(
         """
     #tglogger
@@ -30,7 +38,7 @@ class TelegramFormatter(logging.Formatter):
     *Message*
     {message}
     """
-    ).strip()
+    ).strip()  # noqa
 
     def format(self, record: logging.LogRecord):
         data = dict()
@@ -57,9 +65,9 @@ class TelegramFormatter(logging.Formatter):
             data["zone"] = "No Timezone"
 
         data["level_name"] = record.levelname.lower()
-        data["path"] = reformat_markdown_safe(record.pathname)
+        data["path"] = _reformat_markdown_safe(record.pathname)
         data["lineno"] = record.lineno
-        data["func_name"] = reformat_markdown_safe(record.funcName)
+        data["func_name"] = _reformat_markdown_safe(record.funcName)
         data["thread_id"] = record.thread
         data["thread_name"] = record.threadName