+++ tglogger/handler.py
@@ -1,3 +1,7 @@
+"""
+This module contains a handler for sending log with Telegram.
+"""
+
 import logging
 import os
 import uuid
@@ -17,11 +21,19 @@ logging.setLogRecordFactory(_log_record_uuid)
 
 
 class InvalidBotError(Exception):
+    """
+    This error is raised if the provided bot token is invalid.
+    """
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
 
 class TelegramHandler(logging.Handler):
+    """
+    A `logging.Handler` implementation for sending log records with Telegram.
+    """
+
     def __init__(
         self,
         level=logging.ERROR,
@@ -29,21 +41,19 @@ class TelegramHandler(logging.Handler):
         receiver: str = os.environ["TELEGRAM_RECEIVER"],
         **kwargs
     ):
-        self.bot_token = bot_token
-        self.receiver = receiver
+        self._bot_token = bot_token
+        self._receiver = receiver
         super().__init__(level)
 
         if not kwargs.get("bypass_auth", False):
             self._authorize()
 
     def _authorize(self):
-        url = request._BASE_URL.format(token=self.bot_token, method="getMe")
+        url = request._BASE_URL.format(token=self._bot_token, method="getMe")
         response = request.requests.get(url)
 
         if not response.ok:
-            raise InvalidBotError(
-                "No bot exists with given token. (Hidden for security.)"
-            )
+            raise InvalidBotError("No bot exists with provided token.")
 
     def emit(self, record):
-        return request.send_log(self, record, self.receiver)
+        return request.send_log(self, record, self._receiver)