+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.1a3] - 2020-01-20
+### Changed
+ - Fixed minor bugs.
+
 ## [v0.1.1a2] - 2019-05-26
 ### Added
  - `meta-*.zip` file to log message.